Allow reloading of comments
When a comment is pushed through turbo there is no session present and
therefore there is no current_user.
Without a user context there is no way of telling whether or not the
receiver can edit or delete a comment.

By providing users with a reload button they can rerender the partial
with the correct authorizations.